requirements.txt
numpy==1.26.1
tensorflow==2.20.0
pillow==10.3.0